Use size_t for VertexDataBuffer size instead of uint32_t

+++ PreVEngineTest/prev_test/render/VertexDataBuffer.cpp
@@ -6,7 +6,7 @@ VertexDataBuffer::VertexDataBuffer(const size_t desiredSizeInBytes)
     m_buffer.reserve(desiredSizeInBytes);
 }
 
-void VertexDataBuffer::Add(const void* data, const unsigned int size)
+void VertexDataBuffer::Add(const void* data, const size_t size)
 {
     m_buffer.insert(m_buffer.end(), static_cast<const uint8_t*>(data), static_cast<const uint8_t*>(data) + size);
 }
@@ -21,8 +21,8 @@ const uint8_t* VertexDataBuffer::GetData() const
     return m_buffer.data();
 }
 
-uint32_t VertexDataBuffer::GetSize() const
+size_t VertexDataBuffer::GetSize() const
 {
-    return static_cast<uint32_t>(m_buffer.size());
+    return m_buffer.size();
 }
 } // namespace prev_test::render
